pyproject.toml
[tool.poetry]
name = "anycode"
version = "0.1.0"
description = "Constants and functions generated on-the-fly by OpenAI."
authors = ["Baptiste Fontaine <b@ptistefontaine.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
